Remove waypoint STRIP.STRANECKA.ZHOR from Czech waypoints
The former grass landing strip was turned into an ordinary farming field a few years ago and is no longer guaranteed to be landable (depends on crop type and season). See [1] for an aerial picture. As there are many suitable fields in the vicinity I am removing this obsolete entry.  

[1] https://en.mapy.cz/zakladni?x=15.9317807&y=49.3855429&z=17&l=0&base=ophoto&q=stranecka%20zhor