Add .cup header to all .cup files.
The CUP header defined by Naviter for .cup files is optional for XCSoar.
However Naviter's SeeYou desktop application cannot read a .cup file
which lacks the header.
Therefore add the headers to each .cup file in this repository.
This allows opening all the individual files in SeeYou.
In a recent commit build.sh was enabled to remove the header from individual cup files when it creates the combined global xcsoar_waypoints files